Вдвойне сломали панель накрутки
