+++ Content.Shared/_White/Administration/HoursPanelSystem/HoursPanelMessages.cs
@@ -1,0 +1,24 @@
+using Robust.Shared.Serialization;
+
+namespace Content.Shared._White.Administration;
+
+[Serializable, NetSerializable]
+public sealed class HoursPanelMessageToServer : EntityEventArgs
+{
+    public string PlayerCKey { get; }
+    public string Job { get; }
+    public HoursPanelMessageToServer(string playerCKey, string job)
+    {
+        PlayerCKey = playerCKey;
+        Job = job;
+    }
+}
+
+public sealed class HoursPanelMessageToClient : EntityEventArgs
+{
+    public TimeSpan Time { get; }
+    public HoursPanelMessageToClient(TimeSpan time)
+    {
+        Time = time;
+    }
+}
