--- Content.Shared/_White/Administration/HoursPanelSystem/SharedHoursPanelSystem.cs
@@ -1,18 +1,0 @@
-using Robust.Shared.Network;
-using Robust.Shared.Serialization;
-
-namespace Content.Shared._White.Administration;
-
-public abstract class SharedHoursPanelSystem : EntitySystem
-{
-    public override void Initialize()
-    {
-        base.Initialize();
-
-        SubscribeNetworkEvent<HoursPanelMessage>(OnHoursPanelMessage);
-    }
-
-    protected virtual void OnHoursPanelMessage(HoursPanelMessage message, EntitySessionEventArgs eventArgs)
-    {
-    }
-}
